demo-app/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
opentelemetry-api>=1.24.0
opentelemetry-sdk>=1.24.0
opentelemetry-exporter-otlp-proto-grpc>=1.24.0
opentelemetry-instrumentation-fastapi>=0.45b0
prometheus-fastapi-instrumentator>=7.0.0
prometheus-client>=0.20.0
httpx>=0.27.0
